Can't create a directory tree which does not exist.
This exercise is based on a problem I encountered at work where commons-io was used to
write several files concurrently to the same directory. Initially this directory would
not exist so commons-io will try to create it resulting in an IOException at times.

As commons-io doesn't lend itself well to write a test which reliably proves the problem
I wrote an implementation which mimicks the offending behavior and then wrote an
implementation that gracefully handles directory trees which already exist.